feat: new attestation.id client
Create new attestation ID client with request storage.
This will prevent unnecessary redirects to attestation.id
when authenticating tokens with multiple email addresses.
